The links can have attributes.
For debugging, In the initial Model COFFEE.xml has been added a new symbol "world" linked to the symbol "robot".
The link has geometric attributes that represent the transformation between the "world" and the "robot" and has been labeled as "RT"

The navigation Agent updates the metric information in the symbol "robot" and also in the link "RT".